A
Added basic strucutre for the motif kernel

The idea is to pass the given motifs to the processMotifs function, which converts the list of strings to a trie object.
For each sequence the number of contained motifs is then computed with DFS. The number of motifs in a sequence can be used to compute the similarity between sequences.