Introduced a new folder structure and added a motif and a trie class
Kernel scripts are now in the kernel subfolder. I also added two new
classes. Motif processes a list of strings that are in the same motif
format mentioned in the motif kernel publication. The trie class uses
those motifs to construct a trie. I also implemented a DFS search
function, which can be used to search for motifs in given sequence.