Completed the motif kernel structure
The Motif and Trie classes are now fully working for a specific set of
motifs. The Motif kernel does also work for this set of motifs and
returns a sparse matrix object. I did not yet test any other motif
possible. Also a option to extract motifs from sequences should be
added.