Implemented new package structure
The previous package structure had to be changed in order for the
package to work after installing. Now the package can be installed using
"pip install .". I've only changed the submodule paths for the motif
kernel. Other paths probably have to be adjusted to the new structure.
Additionally I have changed the name of the package to strkernel since
Pyhton packages names cannot include "-".