Corrected library paths and added jupyter notebooks to documentation
I have changed some of the paths used in the mismatch kernel script to
the correct paths for user that installed strkernel. Additionally I have
added the jupyter notebooks to the Tutorial section of the
documentation. Strkernel can now also be installed via pip since if
uploaded it to Pypi.